Persisting Bacon Distance
Because I switched to a node traversal that keeps the entire graph in
memory I can now mark every point on the graph with it's known distance
to KB and use that as a priority heuristic when trying to figure out
which nodes to inspect next. This won't help early in calculating the
graph but should help significantly once there's a large number of
actors and movies calculated.